Remove duplicate benchmarks and minor cleanup
