+++ frontends/benchmarks/typechecker/valid/ListWithSize.scala
@@ -104,34 +104,17 @@ object ListWithSize {
       }
     } ensuring(content(_) == content(l1) ++ content(l2))
 
-    @induct
-    def nilAppend(l : List) : Boolean = {
+    def nilAppend(@induct l : List) : Boolean = {
       append(l, Nil()) == l
     }.holds
 
-    // unclear if we needed this--it was meant to force folding
-    //def appendFold(x : Int, xs : List, ys : List) : Boolean = {
-    //  true
-    //} ensuring (res => res && Cons(x,append(xs, ys)) == append(Cons(x,xs), ys))
+    def appendAssoc(@induct xs : List, ys : List, zs : List) : Boolean = {
+      append(append(xs, ys), zs) == append(xs, append(ys, zs))
+    }.holds
 
-    @induct
-    def appendAssoc(xs : List, ys : List, zs : List) : Boolean = {
-      (append(append(xs, ys), zs) == append(xs, append(ys, zs)))
-    } holds
-
-    def revAuxBroken(l1 : List, e : Int, l2 : List) : Boolean = {
-      (append(reverse(l1), Cons(e,l2)) == reverse0(l1, l2))
-    } // holds
-
-    @induct
-    def reverse0exposed(l1 : List, l2 : List) : Boolean = {
-      (reverse0(l1, l2) == append(reverse(l1), l2))
-    } // holds
-
-    @induct
-    def sizeAppend(l1 : List, l2 : List) : Boolean = {
-      (size(append(l1, l2)) == size(l1) + size(l2))
-    } holds
+    def sizeAppend(@induct l1 : List, l2 : List) : Boolean = {
+      size(append(l1, l2)) == size(l1) + size(l2)
+    }.holds
 
     // proved with unrolling=4
     def concat(l1: List, l2: List) : List = {
@@ -150,13 +133,4 @@ object ListWithSize {
         case Cons(x, xs) => concat0(xs, l2, Cons(x, l3))
       }
     } ensuring(content(_) == content(l1) ++ content(l2) ++ content(l3))
-
-    def reverseConcat0(l1: List, l2: List) : Boolean = {
-      reverse(concat(l1, l2)) == concat(reverse(l2), reverse(l1))
-    } // holds
-
-    // proved with unrolling=2 ???
-    def reverseConcat(l1: List, l2: List) : Boolean = {
-      reverse(concat(l1, l2)) == concat(reverse(l2), reverse(l1))
-    } // holds
 }
